+++ R/fasano.franceschini.test-package.R
@@ -2,7 +2,6 @@
 #' @importFrom Rcpp evalCpp
 #' @importFrom RcppParallel RcppParallelLibs
 #' @importFrom stats binom.test
-#' @importFrom microbenchmark microbenchmark
 #' @useDynLib fasano.franceschini.test, .registration = TRUE
 ## usethis namespace: end
 NULL